Move common Mother test utilities to test module
