+++ stream-chat-android-offline/src/test/java/io/getstream/chat/android/livedata/controller/ChannelControllerImplTest.kt
@@ -17,8 +17,8 @@ import io.getstream.chat.android.client.models.User
 import io.getstream.chat.android.client.utils.Result
 import io.getstream.chat.android.livedata.BaseDomainTest2
 import io.getstream.chat.android.livedata.randomAttachmentsWithFile
-import io.getstream.chat.android.livedata.randomString
 import io.getstream.chat.android.test.TestCall
+import io.getstream.chat.android.test.randomString
 import junit.framework.Assert.assertEquals
 import kotlinx.coroutines.runBlocking
 import org.amshove.kluent.Verify
