+++ stream-chat-android-offline/src/test/java/io/getstream/chat/android/livedata/controller/ThreadControllerImplTest.kt
@@ -10,8 +10,8 @@ import com.nhaarman.mockitokotlin2.whenever
 import io.getstream.chat.android.client.models.Message
 import io.getstream.chat.android.client.utils.Result
 import io.getstream.chat.android.livedata.BaseDomainTest2
-import io.getstream.chat.android.livedata.randomString
 import io.getstream.chat.android.livedata.utils.getOrAwaitValue
+import io.getstream.chat.android.test.randomString
 import kotlinx.coroutines.flow.MutableStateFlow
 import kotlinx.coroutines.test.runBlockingTest
 import org.junit.Before
