+++ stream-chat-android/src/test/java/com/getstream/sdk/chat/viewmodel/messages/MessageListItemLiveDataBenchmark.kt
@@ -4,7 +4,6 @@ import androidx.arch.core.executor.testing.InstantTaskExecutorRule
 import androidx.lifecycle.LiveData
 import androidx.lifecycle.MutableLiveData
 import com.getstream.sdk.chat.adapter.MessageListItem
-import com.getstream.sdk.chat.createDate
 import com.getstream.sdk.chat.createMessage
 import com.getstream.sdk.chat.randomUser
 import com.getstream.sdk.chat.utils.livedata.getOrAwaitValue
@@ -12,6 +11,7 @@ import com.google.common.truth.Truth
 import io.getstream.chat.android.client.models.ChannelUserRead
 import io.getstream.chat.android.client.models.Message
 import io.getstream.chat.android.client.models.User
+import io.getstream.chat.android.test.createDate
 import org.junit.Rule
 import org.junit.Test
 import kotlin.system.measureTimeMillis
