+++ stream-chat-android/src/test/java/com/getstream/sdk/chat/viewmodel/messages/MessageListItemLiveDataTest.kt
@@ -5,7 +5,6 @@ import androidx.lifecycle.LiveData
 import androidx.lifecycle.MutableLiveData
 import androidx.lifecycle.Observer
 import com.getstream.sdk.chat.adapter.MessageListItem
-import com.getstream.sdk.chat.createDate
 import com.getstream.sdk.chat.createMessage
 import com.getstream.sdk.chat.randomUser
 import com.getstream.sdk.chat.utils.livedata.getOrAwaitValue
@@ -17,6 +16,7 @@ import com.nhaarman.mockitokotlin2.verify
 import io.getstream.chat.android.client.models.ChannelUserRead
 import io.getstream.chat.android.client.models.Message
 import io.getstream.chat.android.client.models.User
+import io.getstream.chat.android.test.createDate
 import org.amshove.kluent.any
 import org.junit.Rule
 import org.junit.Test
